refactor(streaming): move agg mod to executor v2
