+++ src/stream/src/executor/hash_agg.rs
@@ -24,8 +24,9 @@ use risingwave_pb::stream_plan;
 use risingwave_pb::stream_plan::stream_node::Node;
 use risingwave_storage::{Keyspace, StateStore};
 
-use super::{AggCall, Executor};
+use super::Executor;
 use crate::executor::{ExecutorBuilder, PkIndices};
+use crate::executor_v2::aggregation::AggCall;
 use crate::executor_v2::{Executor as ExecutorV2, HashAggExecutor};
 use crate::task::{build_agg_call_from_prost, ExecutorParams, LocalStreamManagerCore};
