+++ src/stream/src/executor/integration_tests.rs
@@ -24,6 +24,7 @@ use risingwave_expr::expr::*;
 
 use super::*;
 use crate::executor::test_utils::create_in_memory_keyspace;
+use crate::executor_v2::aggregation::{AggArgs, AggCall};
 use crate::executor_v2::receiver::ReceiverExecutor;
 use crate::executor_v2::{
     Executor as ExecutorV2, LocalSimpleAggExecutor, MergeExecutor, ProjectExecutor,
