+++ src/stream/src/executor/managed_state/aggregation/extreme.rs
@@ -30,7 +30,8 @@ use risingwave_storage::{Keyspace, StateStore};
 
 use super::extreme_serializer::{variants, ExtremePk, ExtremeSerializer};
 use crate::executor::managed_state::flush_status::BtreeMapFlushStatus as FlushStatus;
-use crate::executor::{AggArgs, AggCall, PkDataTypes};
+use crate::executor::PkDataTypes;
+use crate::executor_v2::aggregation::{AggArgs, AggCall};
 
 pub type ManagedMinState<S, A> = GenericExtremeState<S, A, { variants::EXTREME_MIN }>;
 pub type ManagedMaxState<S, A> = GenericExtremeState<S, A, { variants::EXTREME_MAX }>;
