+++ src/stream/src/executor_v2/aggregation/foldable.rs
