+++ src/stream/src/executor_v2/aggregation/row_count.rs
