+++ src/stream/src/executor_v2/hash_agg.rs
@@ -29,11 +29,10 @@ use risingwave_common::util::hash_util::CRC32FastBuilder;
 use risingwave_storage::{Keyspace, StateStore};
 
 use super::{Executor, ExecutorInfo, StreamExecutorResult};
-use crate::executor::{
-    agg_input_arrays, pk_input_arrays, AggCall, AggState, PkDataTypes, PkIndicesRef,
-};
-use crate::executor_v2::agg::{
-    generate_agg_schema, generate_agg_state, AggExecutor, AggExecutorWrapper,
+use crate::executor::{pk_input_arrays, PkDataTypes, PkIndicesRef};
+use crate::executor_v2::aggregation::{
+    agg_input_arrays, generate_agg_schema, generate_agg_state, AggCall, AggExecutor,
+    AggExecutorWrapper, AggState,
 };
 use crate::executor_v2::error::StreamExecutorError;
 use crate::executor_v2::PkIndices;
@@ -395,7 +394,7 @@ mod tests {
     use risingwave_expr::expr::*;
     use risingwave_storage::{Keyspace, StateStore};
 
-    use crate::executor::{AggArgs, AggCall};
+    use crate::executor_v2::aggregation::{AggArgs, AggCall};
     use crate::executor_v2::test_utils::*;
     use crate::executor_v2::{Executor, HashAggExecutor, Message, PkIndices};
     use crate::row_nonnull;
